code related to query separated from model
